Fix get_name_change_details decoding of real API responses (#106)
* Fix NameChangeReviewContext fractional diff fields

The WOM API returns a fractional `hoursDiff` in a name change's review
context (e.g. on a denied `transition_period_too_long` change), but the
model typed `hours_diff`, `ehp_diff`, and `ehb_diff` as `int`, causing
`get_name_change_details` to raise a `msgspec.ValidationError` during
decode. Type these fractional fields as `float`, and align
`negative_gains` values with the sibling `NameChangeData` model (`float`,
to accommodate the `ehp`/`ehb` metrics). `max_hours_diff` and the
total-level fields stay `int`.

* Default Snapshot.imported_at to None for live snapshots

A live (never-imported) player snapshot omits the `importedAt` key
entirely rather than sending it as null. `Snapshot.imported_at` was
`Optional[datetime]` with no default, so msgspec still required the key
to be present and raised `msgspec.ValidationError` when it was absent —
breaking `get_name_change_details` (whose `newStats` is a fresh
hiscores snapshot) and any other snapshot-returning endpoint.

Give `imported_at` a `None` default so an absent key decodes to None,
matching the existing present-but-null behavior. The field moves below
`data`/`created_at` since msgspec requires defaulted fields to follow
non-defaulted ones.

+++ wom/models/names/models.py
@@ -51,7 +51,7 @@ class NameChangeReviewContext(BaseModel):
     # public on the name change endpoints and if we do
     # determine how to handle it in a cleaner way.
 
-    negative_gains: t.Optional[t.Dict[enums.Metric, int]] = None
+    negative_gains: t.Optional[t.Dict[enums.Metric, float]] = None
     """The negative gains that were observed, if there were any. Only populated
     when the reason is `NegativeGains`.
     """
@@ -61,17 +61,17 @@ class NameChangeReviewContext(BaseModel):
     reason is `TransitionTooLong`.
     """
 
-    hours_diff: t.Optional[int] = None
+    hours_diff: t.Optional[float] = None
     """The actual number of hours in the transition period. Only populated when
     reason is `TransitionTooLong` or `ExcessiveGains`.
     """
 
-    ehp_diff: t.Optional[int] = None
+    ehp_diff: t.Optional[float] = None
     """The number difference between the old and new names ehp. Only populated
     when the reason is `ExcessiveGains`.
     """
 
-    ehb_diff: t.Optional[int] = None
+    ehb_diff: t.Optional[float] = None
     """The number difference between the old and new names ehb. Only populated
     when the reason is `ExcessiveGains`.
     """
